reorg state stream functions
