+++ packages/client/src/engine/encoders/index.ts
@@ -1,2 +1,4 @@
 export { createDecode, createDecoder } from './decode';
 export { createEncoder } from './encode';
+
+export type { Decode } from './decode';
