+++ packages/client/src/workers/sync/stream/index.ts
@@ -1,0 +1,2 @@
+export { connect as connectStreamService } from './connect';
+export { createTransformWorldEvents } from './utils';
